Add Main Activity RecyclerView / Fix partial bug.
Add recyclerview row view: group_item.
Add Group entity.
Add holder and adapter for recyclerview.
Add icon for default profile user's picture.
The recyclerView has no data yet.
Fix a bug that were making the app to make a loop in "onResume()" method between MainActivity and LoginActivity when trying to login, not creating  a  new account. The  new lines introduced are checking if there is  an instance of Authentication that is not null before launch login screen again. This  "fix" introduced a new bug, now when you register a new account the app crashes after signup, but account is succesful created, and can be used for login.