Fix Contacts RecyclerView
RecyclerView in Contacts was not showing the data coming from ChildEventListener () because should be dettached/stopped to make it check again after use it once, not gonna update if there is no new child on the node.
RecyclerView was not working also because the holder was not correctly linked with butterknife, and IDs where duplicated due to copy/paste.